part_of_as0
without needing to copy and paste to see length, the actual word doc length of my part is gonna be lengthier. So if any of you find that you are low on time or length of each section you are tackling, mine will probably be about a page each but we will iron that out sunday. I don't have work sunday so I will be free for almost all of this to be free for this assignment